+++ packages/blocks/core/js/wordpress/post-author-name/index.js
@@ -1,6 +1,11 @@
 // @flow
 
 /**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
  * Blockera dependencies
  */
 import { SharedBlockExtension } from '@blockera/editor';
@@ -15,7 +20,15 @@ export const PostAuthorName: BlockType = {
 	name: 'blockeraPostAuthorName',
 	targetBlock: 'core/post-author-name',
 	blockeraInnerBlocks: {
-		'elements/link': sharedInnerBlocks['elements/link'],
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			label: __('Link', 'blockera'),
+			description: __('Hyperlink element.', 'blockera'),
+			settings: {
+				...sharedInnerBlocks['elements/link'].settings,
+				force: false,
+			},
+		},
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;